home_assistant: fix ntfy and VM integrations
In Home Assistant, pass in the VM credentials in as a token in the auth
header. In Traefik, convert that header into the format for basic auth
and forward to VM. This is needed because HA only integrates with
InfluxDB. VM accepts that protocol but auth is different.

For ntfy, format notify msgs using the rest platform.